pyproject.toml
[tool.poetry]
name = "blog-app"
version = "0.1.0"
description = ""
authors = ["test"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
strawberry-graphql = "^0.258.0"
strawberry-graphql-django = "^0.55.0"
django-debug-toolbar = "^5.0.1"
faker = "^33.3.1"
factory-boy = "^3.3.1"
djangorestframework = "^3.15.2"
django-environ = "^0.12.0"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.1"
pytest = "7.4.0"
pytest-django = "^4.9.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
ignore = ["E127", "E501"]
max-line-length = 88
exclude = [".git", "__pycache__", "venv"]
